Move definitions to cpp file
